Added utils.py.
